requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
anyio=3.3.4=py38haa244fe_1
argon2-cffi=21.1.0=py38h294d835_2
async-timeout=4.0.2=pypi_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifultable=1.0.1=pypi_0
bleach=4.1.0=pyhd8ed1ab_0
blis=0.7.5=pypi_0
brotlipy=0.7.0=py38h294d835_1003
ca-certificates=2021.10.8=h5b45459_0
catalogue=2.0.6=pypi_0
certifi=2021.10.8=py38haa244fe_1
cffi=1.15.0=py38hd8c33c5_0
chardet=4.0.0=py38haa244fe_2
charset-normalizer=2.0.7=pypi_0
click=7.1.1=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=36.0.0=py38hb7941b4_0
cupy-cuda100=9.6.0=pypi_0
cupy-cuda115=9.6.0=pypi_0
cycler=0.11.0=pypi_0
cymem=2.0.6=pypi_0
datasets=2.2.2.dev0=pypi_0
debugpy=1.5.1=py38h885f38d_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.4=pypi_0
en-core-web-lg=3.2.0=pypi_0
en-core-web-sm=3.0.0=pypi_0
en-core-web-trf=3.0.0=pypi_0
entrypoints=0.3=pyhd8ed1ab_1003
fastrlock=0.8=pypi_0
filelock=3.3.2=pypi_0
fonttools=4.28.2=pypi_0
frozenlist=1.3.0=pypi_0
fsspec=2022.3.0=pypi_0
huggingface-hub=0.6.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.8.2=py38haa244fe_0
importlib_resources=5.4.0=pyhd8ed1ab_0
ipykernel=6.5.0=py38h595d716_1
ipython=7.29.0=py38h595d716_2
ipython_genutils=0.2.0=py_1
jedi=0.18.1=py38haa244fe_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.2.1=pyhd8ed1ab_0
jupyter-http-over-ws=0.0.8=pypi_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.9.1=py38haa244fe_1
jupyter_server=1.12.0=pyhd8ed1ab_0
jupyterlab=3.2.4=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.8.2=pyhd8ed1ab_0
kiwisolver=1.3.2=pypi_0
langcodes=3.3.0=pypi_0
libsodium=1.0.18=h8d14728_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.0.1=py38h294d835_1
matplotlib=3.5.0=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py38h294d835_1005
msys2-conda-epoch=20160418=1
multidict=6.0.2=pypi_0
multiprocess=0.70.12.2=pypi_0
murmurhash=1.0.6=pypi_0
nbclassic=0.3.4=pyhd8ed1ab_0
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.3.0=py38haa244fe_1
nbformat=5.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.1=pyhd8ed1ab_0
notebook=6.4.6=pyha770c72_0
numpy=1.21.4=pypi_0
openssl=1.1.1l=h8ffe710_0
packaging=21.2=pypi_0
pandas=1.4.2=pypi_0
pandoc=2.16.1=h8ffe710_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.2=pyhd8ed1ab_0
pathlib=1.0.1=pypi_0
pathy=0.6.1=pypi_0
pickleshare=0.7.5=py_1003
pillow=8.4.0=pypi_0
pip=21.0.1=py38haa95532_0
preshed=3.0.6=pypi_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.22=pyha770c72_0
pyarrow=8.0.0=pypi_0
pyconll=3.1.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.7.4=pypi_0
pygments=2.10.0=pyhd8ed1ab_0
pyopenssl=21.0.0=pyhd8ed1ab_0
pyparsing=2.4.7=pypi_0
pyrsistent=0.18.0=py38h294d835_0
pysocks=1.7.1=py38haa244fe_4
python=3.8.12=h6244533_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd8ed1ab_0
pywin32=302=py38h294d835_2
pywinpty=0.5.7=py38_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py38h09162b1_1
regex=2021.11.10=pypi_0
requests=2.26.0=pyhd8ed1ab_0
responses=0.18.0=pypi_0
sacremoses=0.0.46=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.2=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py38haa95532_0
setuptools-scm=6.3.2=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=5.2.1=pypi_0
sniffio=1.2.0=py38haa244fe_2
spacy=3.2.0=pypi_0
spacy-alignments=0.8.4=pypi_0
spacy-legacy=3.0.8=pypi_0
spacy-loggers=1.0.1=pypi_0
spacy-transformers=1.0.4=pypi_0
sqlite=3.36.0=h2bbff1b_0
srsly=2.4.2=pypi_0
terminado=0.9.4=py38haa95532_0
testpath=0.5.0=pyhd8ed1ab_0
thinc=8.0.13=pypi_0
threadpoolctl=3.0.0=pypi_0
tokenizers=0.10.3=pypi_0
tomli=1.2.2=pypi_0
torch=1.10.0+cu113=pypi_0
torchaudio=0.10.0+cu113=pypi_0
torchvision=0.11.1+cu113=pypi_0
tornado=6.1=py38h294d835_2
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.9.2=pypi_0
typer=0.3.2=pypi_0
typing-extensions=3.10.0.2=pypi_0
urllib3=1.26.7=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wasabi=0.8.2=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
websocket-client=1.2.1=py38haa244fe_0
wheel=0.37.0=pyhd3eb1b0_1
win_inet_pton=1.1.0=py38haa244fe_3
wincertstore=0.2=py38haa95532_2
winpty=0.4.3=4
xxhash=3.0.0=pypi_0
yarl=1.7.2=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.6.0=pyhd8ed1ab_0
